Autotests/requirements.txt
pytest
requests
python-dotenv~=1.0.0
urllib3